PiperOrigin-RevId: 741272687

+++ minimal_task_runner.py
@@ -108,7 +108,7 @@ def _main() -> None:
 
   print('Goal: ' + str(task.goal))
   is_done = False
-  for _ in range(task.complexity * 10):
+  for _ in range(int(task.complexity * 10)):
     response = agent.step(task.goal)
     if response.done:
       is_done = True
